unify-criteria: Added code to delete criteria before uploading new ones.
